+++ tasmota/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 7.1.2.6 20191214
 
 - Change some more Settings locations freeing up space for future single char allowing variable length text
+- Change tasmota-basic.bin and FIRMWARE_BASIC to tasmota-lite.bin and FIRMWARE_LITE
 - Add Zigbee send automatic ZigbeeRead after sending a command
 - Add Zigbee improving Occupancy:false detection for Aqara sensor
 - Add fallback functionality from next version 8.x